accounts model update

+++ app/django/accounts/models.py
@@ -1,15 +1,13 @@
 import os
 
-import magic
-from django.conf import settings
 from django.contrib.auth.base_user import AbstractBaseUser
 from django.contrib.auth.models import Group
 from django.contrib.auth.models import UserManager, PermissionsMixin
 from django.contrib.auth.validators import UnicodeUsernameValidator
 from django.core.mail import send_mail
 from django.db import models
-from django.db.models.signals import pre_delete
 from django.dispatch import receiver
+from django.db.models.signals import pre_save, post_delete
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
@@ -133,7 +131,7 @@ class Profile(models.Model):
     name = models.CharField('성명', max_length=20, blank=True)
     birth_date = models.DateField('생년월일', null=True, blank=True)
     cell_phone = models.CharField('휴대폰', max_length=13, blank=True)
-    # image = models.ImageField(upload_to='users/', null=True, blank=True, verbose_name='프로필 이미지')
+    image = models.ImageField(upload_to='users/', null=True, blank=True, verbose_name='프로필 이미지')
     like_posts = models.ManyToManyField(Post, blank=True, related_name='post_likes')
     like_comments = models.ManyToManyField(Comment, blank=True, related_name='comment_likes')
     blame_posts = models.ManyToManyField(Post, blank=True, related_name='post_blames')
@@ -148,32 +146,26 @@ class Profile(models.Model):
         verbose_name_plural = '사용자 프로필'
 
 
-class Image(models.Model):
-    profile = models.OneToOneField(Profile, on_delete=models.CASCADE, verbose_name='프로필')
-    image = models.ImageField(upload_to='users/', verbose_name='프로필 이미지')
-    image_name = models.CharField('파일명', max_length=100, blank=True)
-    image_type = models.CharField('타입', max_length=100, blank=True)
-    image_size = models.PositiveBigIntegerField('사이즈', blank=True, null=True)
-    created = models.DateTimeField(auto_now_add=True)
+@receiver(pre_save, sender=Profile)
+def delete_old_image(sender, instance, **kwargs):
+    if not instance.pk:
+        return False  # 신규 객체일 경우 무시
 
-    def __str__(self):
-        return settings.MEDIA_URL
+    try:
+        old_image = sender.objects.get(pk=instance.pk).image
+    except sender.DoesNotExist:
+        return False
 
-    def save(self, *args, **kwargs):
-        if self.image:
-            self.image_name = self.image.name.split('/')[-1]
-            mime = magic.Magic(mime=True)
-            self.image_type = mime.from_buffer(self.image.read())
-            self.image_size = self.image.size
-        super().save(*args, **kwargs)
+    new_image = instance.image
+    if not old_image == new_image:
+        if old_image and os.path.isfile(old_image.path):
+            os.remove(old_image.path)
 
 
-@receiver(pre_delete, sender=Image)
-def delete_file_on_delete(sender, instance, **kwargs):
-    # Check if the file exists before attempting to delete it
-    if instance.image:
-        if os.path.isfile(instance.image.path):
-            os.remove(instance.image.path)
+@receiver(post_delete, sender=Profile)
+def delete_image_on_delete(sender, instance, **kwargs):
+    if instance.image and os.path.isfile(instance.image.path):
+        os.remove(instance.image.path)
 
 
 class Scrape(models.Model):
